requirements.txt
numpy >= 1.21.4
matplotlib >= 3.5.0
seaborn >= 0.11.2
torchtext >= 0.11.0
scikit-learn >= 1.0.1
transformers >= 4.12.5
captum >= 0.4.1
pandas >= 1.3.4
nltk >= 3.6.5
spacy >= 3.2.1
PyYAML >= 6.0
emoji >= 1.6.1
tqdm >= 4.47.0
gensim >= 4.1.2
pyLDAvis >= 3.3.1
wordcloud == 1.8.1